+++ packages/koot-webpack/plugins/spa-template.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-console */
-// https://github.com/jantimon/html-webpack-plugin
+// ref: https://github.com/jantimon/html-webpack-plugin
 
 const fs = require('fs-extra');
 const path = require('path');