+++ test/projects/standard/src/views/static/index.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import { extend } from 'koot';
 
+import photo from '@assets/photo.jpg';
+
 export default extend({
     pageinfo: (/*state, renderProps*/) => ({
         title: `${__('pages.static.title')} - ${__('title')}`,
@@ -18,7 +20,7 @@ export default extend({
             <div className="block">
                 <h2>{__('pages.static.method_require')}</h2>
                 <p>{__('pages.static.method_require_content')}</p>
-                <img src={require('@assets/photo.jpg')} />
+                <img src={photo} />
             </div>
             <div className="block">
                 <h2>{__('pages.static.method_static')}</h2>